apps/backend/requirements.txt
# Core LangChain dependencies
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-google-genai>=2.0,<3.0
langgraph>=0.2,<0.3

# Vector store for memory
faiss-cpu>=1.7.4

sentence-transformers>=2.2.2

# PersonaPlex compatible versions (torch 2.4.x)
torch>=2.9.0
huggingface-hub>=0.24.0
transformers>=4.40.0

# Vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
pypdf>=4.0.0

# Web search
requests>=2.31.0
httpx>=0.27.0

youtube-transcript-api>=0.6.2

playwright>=1.41.0

pywinauto>=0.6.8; platform_system == "Windows"
pyautogui>=0.9.54

tool-calling-llm>=0.1.2

# API server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.5.0

# Logging
loguru>=0.7.0

# Optional: Local LLM support
langchain-ollama>=0.3,<0.4
llama-cpp-python>=0.2.0; python_version < "3.14"

# Utilities
python-dateutil>=2.8.2
mss>=9.0.0

# PersonaPlex streaming audio
websockets>=12.0
sounddevice>=0.4.6
sphn>=0.1.0
numpy>=1.24.0

# Discord bot
discord.py>=2.3.0

# Heartbeat cron scheduler
croniter>=1.3.8

# Google Calendar integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Spotify integration
spotipy>=2.23.0

# Notion integration
notion-client>=2.0.0

# GitHub integration
PyGithub>=2.1.0
